pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bcrypt>=4.3.0",
    "email-validator>=2.2.0",
    "fastapi>=0.115.11",
    "nltk>=3.9.1",
    "pydantic>=2.10.6",
    "python-jose>=3.4.0",
    "python-multipart>=0.0.20",
    "speechrecognition>=3.14.2",
    "sqlalchemy>=2.0.39",
    "uvicorn>=0.34.0",
]
